# packages/api-contracts/client/v1/openapi.yaml
openapi: 3.1.0
info:
  title: KMS TV — Client API
  version: 1.0.0
  description: |
    The viewer-facing surface, called by the web, mobile and television
    applications.

    **This document is the contract, not a description of the implementation**
    (ADR-0010). The API changes here first, in a reviewed pull request; the
    server is then verified against it in both directions by
    `tests/Feature/Contract/OpenApiConformanceTest.php`.

    Within major version 1 this surface is **append-only**. A Smart TV
    application reaches its installed base over months to years and some devices
    effectively stop updating, so a breaking change is not a bug to fix forward —
    it is a population of devices that stops working. See docs/api/versioning.md.
  license:
    name: Proprietary
servers:
  - url: /api/client/v1
...
paths:
...
  /subscription:
    get:
      tags: [Commerce]
      operationId: getSubscription
      summary: The account's live subscription, if any
      responses:
        '200':
          description: Subscription or null
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { oneOf: [{ $ref: '#/components/schemas/Subscription' }, { type: 'null' }] }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
    post:
      tags: [Commerce]
      operationId: subscribe
      summary: Start a subscription
      description: |
        **No payment is taken.** No payment provider has been selected, and none
        is integrated in this phase — the subscription and the entitlement it
        produces are real, the money movement is not yet built.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [plan_id]
              properties:
                plan_id: { $ref: '#/components/schemas/Identifier' }
      responses:
        '201':
          description: Subscribed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Subscription' }
        '401': { $ref: '#/components/responses/Problem' }
        '404': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A short-lived, device-bound access token. Account, profile and device
        identity all come from the token — never from a request body, because a
        client-asserted device identifier would let a caller inherit another
        device's class, which is an input to the rights decision from Phase 4.
  responses:
    Problem:
      description: RFC 9457 problem document
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Subscription:
      type: object
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        plan_id: { $ref: '#/components/schemas/Identifier' }
        status: { type: string, enum: [trialing, active, past_due, cancelled, expired] }
        started_at: { type: [string, 'null'], format: date-time }
        current_period_end: { type: [string, 'null'], format: date-time }
        trial_ends_at: { type: [string, 'null'], format: date-time }
        cancel_at_period_end: { type: boolean }
...
    Identifier:
      type: string
      format: uuid
      description: UUIDv7. Sequential integers are never exposed.
      examples: ['0199c1a4-7e3b-7c21-9f4d-2b6a8e05d913']
    Problem:
      type: object
      required: [type, title, status, code, correlation_id]
      properties:
        type: { type: string, format: uri }
        title:
          type: string
          description: Non-localised summary. Clients localise from the code, never from this.
        status: { type: integer }
        code:
          type: string
          description: |
            Stable machine-readable code. **Clients switch on this and never on
            prose.** A code's meaning never changes once shipped; the registry is
            append-only.
        detail: { type: string }
        instance: { type: string }
        correlation_id: { type: string }
        meta: { type: object, additionalProperties: true }
...
security:
  - bearerAuth: []